Changes to ECS, completed deserialisation
Added two new functions for ECS:
-> HasComponent:
- Before, the ECS structure will throw an error if no component is found within the system. Added this function as a utility for users to check if an entity has a component instead of throwing an error.
- Returns true if the Entity has the component, else false

-> GetEntities
- Return a list of active entities in the scene

-> Completed the required feature of loading data from file during load time  using Darrion's serializer

+++ Artless/ALEngine/Source/ECS/Systems/RenderSystem.cpp
@@ -56,7 +56,7 @@ namespace ALEngine::ECS
 			{  0.5f, -0.5f }	// btm right
 		};
 
-		u64 constexpr INDICES_SIZE{ 6 };
+		s32 constexpr INDICES_SIZE{ 6 };
 	}
 
 	void RenderSystem::RenderBatch(void)
